MB-26549 Set wait time for bad config to 2s
A bad configurations returned from the server, it indicates that
the bucket is being setup. When that happens, it makes sense to
have a different waiting time between attempts. Currently, the
default wait time between attempts is 500ms, with 5 attempts total.
This gives a total of 2s when attempting to connect a bucket that
is not ready. This time is not enough and causes sporadic test
failures. For this reason, the waiting time between attempts for
the bad configuration case is hard coded as 2s.

Change-Id: Ie3f2149dd554c6cc4483fc14d62c2ba9bb8306b7
Reviewed-on: http://review.couchbase.org/85203
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>