MB-30267 Adjust start snapshot value when re-opening a stream
For an idle open connection, if a new mutation is received,
a new snapshot marker is sent to the consumer of DCP. The
marker will have the start and end values be the same and both
equals to the last seen sequence number + 1.
If the connection drops at that point, then an attempt
to re-open the connection would fail if the client use these
values for start and end snapshot values.
To fix that, if the sequence number last seen is smaller than
the start seuqnece number, then we will lower the start sequence
number.

Change-Id: Ib7c777c9f0c98389089d82dee952b3f48fec9312
Reviewed-on: http://review.couchbase.org/96423
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Till Westmann <till@couchbase.com>