MB-56436: update seqno in partition state, even when it goes backwards
Under normal circumstances, the seqno being streamed from DCP is ever
increasing, however should KV lose the cursor and be unable to resume
it at the current position, it will end the stream will restart the
stream at the start of the snapshot. In this case, the first seqno
received will be lower than was previously received. While we
will continue to issue a warning in this case, we will then accept the
seqno and not issue further warnings for any given vbucket rollback.

Change-Id: Ie504a377a311168714425078baa2b04a76cba87e
Reviewed-on: https://review.couchbase.org/c/analytics-dcp-client/+/189683
Tested-by: Michael Blow <michael.blow@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>
Well-Formed: Restriction Checker