MB-63488: Disregard bucket NOT_FOUND returned by memcached
Details:
If memcached is asked about buckets while it is warming up, it returns bucket
NOT_FOUND even if the bucket exists. We originally throw BucketNotFoundException
which indicates that we should stop retrying to connect and rollback the collections
on that bucket. This change disregards the bucket NOT_FOUND returned
by memcached and instead uses ns_server to decide whether a bucket exists or not.

Change-Id: I78688c01247e94c57490fed1281c010fc9923d49
Reviewed-on: https://review.couchbase.org/c/analytics-dcp-client/+/216305
Tested-by: Hussain Towaileb <Hussain.Towaileb@Couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Murtadha Hubail <Murtadha.hubail@couchbase.com>